// frontend/src/openapi/model/fichajeEditRequestDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface FichajeEditRequestDto { 
    id: number;
    trabajadorId: number;
    empresaId: number;
    fechaInicio: string;
    fechaFin?: string;
}