// frontend/src/openapi/model/fichajesDto.ts
/**
 * OpenAPI definition
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { EmpresaDto } from './empresaDto';
import { TrabajadorDto } from './trabajadorDto';


export interface FichajesDto { 
    id?: number;
    trabajador?: TrabajadorDto;
    empresa?: EmpresaDto;
    fechaInicio?: string;
    fechaFin?: string;
}